define SchemaTree
