+++ server/src/main/java/org/apache/iotdb/db/mpp/common/schematree/PathPatternTree.java
@@ -17,12 +17,27 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.sql.analyze;
+package org.apache.iotdb.db.mpp.common.schematree;
 
-import java.util.List;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.ByteBuffer;
 
-public class PathPatternNode {
+public class PathPatternTree {
 
-  private String name;
-  private List<PathPatternNode> childs;
+  private PathPatternNode root;
+
+  /**
+   * Since IoTDB v0.13, all DDL and DML use patternMatch as default. Before IoTDB v0.13, all DDL and
+   * DML use prefixMatch.
+   */
+  protected boolean isPrefixMatchPath;
+
+  public void serialize(OutputStream baos) throws IOException {
+    // TODO
+  }
+
+  public void deserialize(ByteBuffer buffer) throws IOException {
+    // TODO
+  }
 }
