+++ server/src/main/java/org/apache/iotdb/db/mpp/common/schematree/SchemaTree.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -16,26 +16,20 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.mpp.sql.analyze;
 
-import org.apache.iotdb.db.metadata.path.PartialPath;
-import org.apache.iotdb.db.mpp.common.schematree.PathPatternTree;
-import org.apache.iotdb.db.mpp.common.schematree.SchemaTree;
-import org.apache.iotdb.tsfile.write.schema.MeasurementSchema;
+package org.apache.iotdb.db.mpp.common.schematree;
 
-import java.util.List;
-import java.util.Map;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.ByteBuffer;
 
-public class ClusterSchemaFetcher implements ISchemaFetcher {
+public class SchemaTree {
 
-  @Override
-  public Map<String, MeasurementSchema> fetchSchema(
-      PartialPath deviceId, List<String> measurementIdList) {
-    return null;
+  public void serialize(OutputStream baos) throws IOException {
+    // TODO
   }
 
-  @Override
-  public SchemaTree fetchSchema(PathPatternTree patternTree) {
-    return null;
+  public void deserialize(ByteBuffer buffer) throws IOException {
+    // TODO
   }
 }
